requirements.txt
numpy>=1.23
pillow>=10.0
tqdm>=4.66
pandas>=2.0
matplotlib>=3.8

torch>=2.1
torchvision>=0.16

pywebview>=5.1
